Extract crumb stack module from browse view
views/browse.js was 1185 LOC and embedded a coherent crumb-stack module:
parts value type, token parser/emitter, pill-bar + trail renderer, filter
badges, and the async label hydration that upgrades raw tokens to
human-readable names. Half of those internals were already exported just
so test_browse_crumbs.js could reach them.

Move all of that to views/browse/crumbs.js — the sole owner of the parts
value type and the crumb stack. views/browse.js keeps mounting (root tabs,
drill frame, show-landing dispatch), the drill loader, the filter input,
and imports the rest back from crumbs.js.

browse.js: 1185 -> 520 LOC. No duplicate definitions remain.
